requirements.txt
pygments==2.4.1
nlpaug
numpy
python-dotenv
mysql-connector
opennmt-py==1.2.0
jupyter
matplotlib
networkx
nltk>=3.4.5
transformers>=2.5.0
flask
sklearn
networkx
matplotlib

